arm64: Handle Bluetooth on the Pinebook Pro via hci_uart
This removed the need for handling Bluetooth via hciattach.
However, the ap6255 seems to be a little quirky. It does
not provide a unique MAC address. Thus userspace needs to
set one via `btmgmt --index hci1 public-addr 43:43:a0:12:1f:aa`
or a compareable tool that can write to the HCI control management
channel.