arm64/dts: Add manual speaker volume control
This commit adds an ALSA control for the speaḱer amplifier.

Whiel this gives pulseaudio a meaningful distinction between speaker and headphones
pulseaudio does unfortunately also turn down the headphones by -48 dB whenever
the speaker is selected. Since the headphone out and the speaker in are connected
in parallel this reduces maximum volume of the speakers drastically. However, I deem
it better than not being able to control the amplifier state at all.